requirements.txt
azure-core==1.26.0
azure-identity==1.11.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
idna==3.4
msal==1.20.0
msal-extensions==1.0.0
msgraph-core==0.2.2
portalocker==2.6.0
pycparser==2.21
PyJWT==2.6.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
typing_extensions==4.4.0
urllib3==1.26.12